+++ detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/console/BuildFailureReportSpec.kt
@@ -33,25 +33,8 @@ internal class BuildFailureReportSpec : Spek({
                 assertThat(report).isNull()
             }
 
-            it("should print a warning in yellow if threshold met") {
-                subject.init(TestConfig(mapOf("warningThreshold" to "1")))
-                val report = subject.render(detektion)
-                val expectedMessage = "Warning: 1 weighted code smells found. Warning threshold is 1 and fail threshold is -1!"
-                assertThat(report).isEqualTo(expectedMessage.yellow())
-            }
-
-            it("should throw a build failure error") {
-                subject.init(TestConfig(mapOf("failThreshold" to "-2")))
-                assertThatExceptionOfType(BuildFailure::class.java).isThrownBy { subject.render(detektion) }
-            }
-
             it("should throw a build failure error when maxIssues met") {
                 subject.init(TestConfig(mapOf("maxIssues" to "-2")))
-                assertThatExceptionOfType(BuildFailure::class.java).isThrownBy { subject.render(detektion) }
-            }
-
-            it("should throw a build failure error even if warning threshold is also met") {
-                subject.init(TestConfig(mapOf("failThreshold" to "-2", "warningThreshold" to "-2")))
                 assertThatExceptionOfType(BuildFailure::class.java).isThrownBy { subject.render(detektion) }
             }
 
@@ -63,6 +46,19 @@ internal class BuildFailureReportSpec : Spek({
                 } catch (e: BuildFailure) {
                     assertThat(e.stackTrace).isEmpty()
                 }
+            }
+
+            it("should print a warning in yellow if weighted issues are not zero but below threshold") {
+                subject.init(TestConfig(mapOf("maxIssues" to "10")))
+                val report = subject.render(detektion)
+                val expectedMessage = "Build succeeded with 1 weighted issues (threshold defined was 10)."
+                assertThat(report).isEqualTo(expectedMessage.yellow())
+            }
+
+            it("should not print a warning if weighted issues are zero") {
+                subject.init(TestConfig(mapOf("maxIssues" to "10")))
+                val report = subject.render(TestDetektion())
+                assertThat(report).isNull()
             }
         }
     }