+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/style/UnusedImports.kt
@@ -20,6 +20,8 @@ import org.jetbrains.kotlin.psi.KtReferenceExpression
 
 /**
  * This rule reports unused imports. Unused imports are dead code and should be removed.
+ * Exempt from this rule are imports resulting from references to elements within KDoc and
+ * from destructuring declarations (componentN imports).
  *
  * @author Artur Bosch
  * @author Marvin Ramin
@@ -41,6 +43,7 @@ class UnusedImports(config: Config) : Rule(config) {
         private val kotlinDocReferencesRegExp = Regex("\\[([^]]+)](?!\\[)")
         private val kotlinDocSeeReferenceRegExp = Regex("^@see (.+)")
         private val whiteSpaceRegex = Regex("\\s+")
+        private val componentNRegex = Regex("component\\d+")
     }
 
     override fun visit(root: KtFile) {
@@ -78,6 +81,7 @@ class UnusedImports(config: Config) : Rule(config) {
                     .filter { it.identifier()?.contains("*")?.not() == true }
                     .filter { it.identifier() != null }
                     .filter { !operatorSet.contains(it.identifier()) }
+                    .filter { !componentNRegex.matches(it.identifier()!!) }
             super.visitImportList(importList)
         }