+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/style/UseRequire.kt
@@ -8,6 +8,7 @@ import io.gitlab.arturbosch.detekt.api.Issue
 import io.gitlab.arturbosch.detekt.api.Rule
 import io.gitlab.arturbosch.detekt.api.Severity
 import io.gitlab.arturbosch.detekt.rules.argumentCount
+import io.gitlab.arturbosch.detekt.rules.isEnclosedByConditionalStatement
 import io.gitlab.arturbosch.detekt.rules.isIllegalArgumentException
 import org.jetbrains.kotlin.psi.KtBlockExpression
 import org.jetbrains.kotlin.psi.KtNamedFunction
@@ -38,9 +39,12 @@ class UseRequire(config: Config = Config.empty) : Rule(config) {
     )
 
     override fun visitThrowExpression(expression: KtThrowExpression) {
+        if (!expression.isIllegalArgumentException()) return
+
         if (expression.isOnlyExpressionInBlock()) return
 
-        if (expression.isIllegalArgumentException() && expression.argumentCount < 2) {
+        if (expression.isEnclosedByConditionalStatement() &&
+            expression.argumentCount < 2) {
             report(CodeSmell(issue, Entity.from(expression), issue.description))
         }
     }