+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/exceptions/ThrowingExceptionInMainSpec.kt
@@ -1,5 +1,7 @@
 package io.gitlab.arturbosch.detekt.rules.exceptions
 
+import io.gitlab.arturbosch.detekt.test.compileAndLint
+import io.gitlab.arturbosch.detekt.test.compileContentForTest
 import io.gitlab.arturbosch.detekt.test.lint
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
@@ -11,17 +13,19 @@ class ThrowingExceptionInMainSpec : Spek({
     describe("ThrowingExceptionInMain rule") {
 
         it("has a runnable main method which throws an exception") {
-            val code = "fun main(args: Array<String>) { throw new IOException() }"
-            assertThat(subject.lint(code)).hasSize(1)
+            val code = "fun main(args: Array<String>) { throw IllegalArgumentException() }"
+            assertThat(subject.compileAndLint(code)).hasSize(1)
         }
 
         it("has wrong main methods") {
-            val code = """
-				fun main(args: Array<String>) { }
-				private fun main() { }
-				fun mai() { }
-				fun main(args: String) { }"""
-            assertThat(subject.lint(code)).hasSize(0)
+            val file = compileContentForTest(
+                """
+                    fun main(args: Array<String>) { }
+                    private fun main() { }
+                    fun mai() { }
+                    fun main(args: String) { }"""
+            )
+            assertThat(subject.lint(file)).hasSize(0)
         }
     }
 })